backend/requirements.txt
django
djangorestframework
requests
djangorestframework-simplejwt
drf_yasg
gunicorn
django-cors-headers
whitenoise